third_party/WebKit/LayoutTests/external/wpt/service-workers/service-worker/multi-globals/url-parsing.https-expected.txt

This is a testharness.js-based test.
FAIL register should use the relevant global of the object it was called on to resolve the script URL assert_equals: the service worker found at relevant/test-sw.js must have been the one to intercept the fetch expected "relevant" but got "entry"
FAIL register should use the relevant global of the object it was called on to resolve the scope URL assert_equals: the scope URL should be relevant/scope expected "https://web-platform.test:8444/service-workers/service-worker/multi-globals/relevant/scope" but got "https://web-platform.test:8444/service-workers/service-worker/multi-globals/current/scope"
FAIL getRegistration should use the relevant global of the object it was called on to resolve the script URL assert_equals: the retrieved registration's scope should be equal to the original's scope expected "https://web-platform.test:8444/service-workers/service-worker/multi-globals/relevant/" but got "https://web-platform.test:8444/service-workers/service-worker/multi-globals/"
Harness: the test ran to completion.
